3/26 3:29
Whoops! I just realized that I was doing it in postfix when I should have been doing it in prefix. I've changed some of the commands to now wait for when there are 2 numbers in the stack and then pop the top operation. It turns out I needed an operation stack after all